requirements.txt
dagshub==0.3.12
deepchecks==0.17.5
dvc==3.37.0
fastapi==0.108.0
hydra-core==1.3.2
ipykernel==6.28.0
joblib==1.3.2
json5==0.9.6
jsonschema==4.20.0
matplotlib==3.8.2
mlflow==2.9.2
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
rembg==2.0.52
scikit-learn==1.3.2
streamlit==1.29.0
ultralytics==8.0.227